requirements.txt
asttokens==2.4.1
colorama==0.4.6
executing==2.0.1
icecream==2.1.3
Pillow==10.1.0
Pygments==2.16.1
six==1.16.0
